+++ crates/server/src/store/sqlite/authority.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 use log::{error, info, warn};
 
 use crate::client::op::LowerQuery;
-use crate::client::rr::dnssec::{DnsSecResult, SigSigner, SupportedAlgorithms};
+use crate::client::rr::dnssec::{DnsSecResult, SigSigner};
 use crate::client::rr::{LowerName, RrKey};
 use crate::proto::op::ResponseCode;
 use crate::proto::rr::dnssec::rdata::key::KEY;
@@ -24,7 +24,9 @@ use crate::proto::rr::{DNSClass, Name, RData, Record, RecordSet, RecordType};
 
 #[cfg(feature = "dnssec")]
 use crate::authority::UpdateRequest;
-use crate::authority::{Authority, LookupError, MessageRequest, UpdateResult, ZoneType};
+use crate::authority::{
+    Authority, LookupError, LookupOptions, MessageRequest, UpdateResult, ZoneType,
+};
 use crate::error::{PersistenceErrorKind, PersistenceResult};
 use crate::store::in_memory::InMemoryAuthority;
 use crate::store::sqlite::{Journal, SqliteConfig};
@@ -313,8 +315,7 @@ impl SqliteAuthority {
                                 if block_on(self.lookup(
                                     &required_name,
                                     RecordType::ANY,
-                                    false,
-                                    SupportedAlgorithms::new(),
+                                    LookupOptions::default(),
                                 ))
                                 .unwrap_or_default()
                                 .was_empty()
@@ -330,8 +331,7 @@ impl SqliteAuthority {
                                 if block_on(self.lookup(
                                     &required_name,
                                     rrset,
-                                    false,
-                                    SupportedAlgorithms::new(),
+                                    LookupOptions::default(),
                                 ))
                                 .unwrap_or_default()
                                 .was_empty()
@@ -355,8 +355,7 @@ impl SqliteAuthority {
                                 if !block_on(self.lookup(
                                     &required_name,
                                     RecordType::ANY,
-                                    false,
-                                    SupportedAlgorithms::new(),
+                                    LookupOptions::default(),
                                 ))
                                 .unwrap_or_default()
                                 .was_empty()
@@ -372,8 +371,7 @@ impl SqliteAuthority {
                                 if !block_on(self.lookup(
                                     &required_name,
                                     rrset,
-                                    false,
-                                    SupportedAlgorithms::new(),
+                                    LookupOptions::default(),
                                 ))
                                 .unwrap_or_default()
                                 .was_empty()
@@ -395,8 +393,7 @@ impl SqliteAuthority {
                     if !block_on(self.lookup(
                         &required_name,
                         require.rr_type(),
-                        false,
-                        SupportedAlgorithms::new(),
+                        LookupOptions::default(),
                     ))
                     .unwrap_or_default()
                     .iter()
@@ -482,12 +479,8 @@ impl SqliteAuthority {
                 .any(|sig| {
                     let name = LowerName::from(sig.signer_name());
                     // TODO: updates should be async as well.
-                    let keys = block_on(self.lookup(
-                        &name,
-                        RecordType::KEY,
-                        false,
-                        SupportedAlgorithms::new(),
-                    ));
+                    let keys =
+                        block_on(self.lookup(&name, RecordType::KEY, LookupOptions::default()));
 
                     let keys = match keys {
                         Ok(keys) => keys,
@@ -940,21 +933,17 @@ impl Authority for SqliteAuthority {
         &self,
         name: &LowerName,
         rtype: RecordType,
-        is_secure: bool,
-        supported_algorithms: SupportedAlgorithms,
+        lookup_options: LookupOptions,
     ) -> Pin<Box<dyn Future<Output = Result<Self::Lookup, LookupError>> + Send>> {
-        self.in_memory
-            .lookup(name, rtype, is_secure, supported_algorithms)
+        self.in_memory.lookup(name, rtype, lookup_options)
     }
 
     fn search(
         &self,
         query: &LowerQuery,
-        is_secure: bool,
-        supported_algorithms: SupportedAlgorithms,
+        lookup_options: LookupOptions,
     ) -> Pin<Box<dyn Future<Output = Result<Self::Lookup, LookupError>> + Send>> {
-        self.in_memory
-            .search(query, is_secure, supported_algorithms)
+        self.in_memory.search(query, lookup_options)
     }
 
     /// Return the NSEC records based on the given name
@@ -967,11 +956,9 @@ impl Authority for SqliteAuthority {
     fn get_nsec_records(
         &self,
         name: &LowerName,
-        is_secure: bool,
-        supported_algorithms: SupportedAlgorithms,
+        lookup_options: LookupOptions,
     ) -> Pin<Box<dyn Future<Output = Result<Self::Lookup, LookupError>> + Send>> {
-        self.in_memory
-            .get_nsec_records(name, is_secure, supported_algorithms)
+        self.in_memory.get_nsec_records(name, lookup_options)
     }
 
     fn add_update_auth_key(&mut self, name: Name, key: KEY) -> DnsSecResult<()> {